Quote boolean values
This is helpful because Postgres returns the strings 't' and 'f' for values in a boolean column, but the (unquoted) literals t and f are not acceptable in queries. However, the quoted strings 't' and 'f' are accepted (see http://www.postgresql.org/docs/9.1/static/datatype-boolean.html), so this change means that you can use the value returned from Postgres in a query without needing to transform it.