Optimised ensure_inserted. Fixed bug with insert_or_update.
Ensure inserted now inserts and obtains the ID in a single query.

If one used the same field in both the condition and the update tuples in
insert_or_update, an error was thrown due to incorrect SQL syntax (double
specification of a field). This has now been fixed: fields in condition
tuples also found in update tuples are now ignored.

Signed-off-by: Rok Strniša <rok.strnisa@citrix.com>